move SECRET_PLACEHOLDER to avoid circular deps
