+++ tests/functional/dependencies/test_dependency_secrets.py
@@ -25,7 +25,7 @@ class TestAllowSecretProfilePackage:
 
     def test_allow_secrets(self, project):
         _, log_output = run_dbt_and_capture(["deps"])
-        # this will not be written to logs or lock file
+        # this will not be written to logs
         assert not ("super secret" in log_output)
         assert "*****" in log_output
         assert not ("DBT_ENV_SECRET_FOR_LOGGING" in log_output)
